simplify did resolver
