+++ crates/datasec/src/did.rs
@@ -23,63 +23,50 @@ use crate::{did, DidResolver};
 /// Did Result type using did module-specific Error
 pub type Result<T> = std::result::Result<T, Error>;
 
-/// Resolve a DID to a DID document.
-///
-/// The [DID resolution](https://www.w3.org/TR/did-core/#did-resolution) functions
-/// resolve a DID into a DID document by using the "Read" operation of the applicable
-/// DID method.
-///
-/// Caveats:
-/// - No JSON-LD Processing, however, valid JSON-LD is returned.
-/// - Ignores accept header.
-/// - Only returns application/did+ld+json.
-/// - did:key support for ed25519
-/// - did:web support for .well-known and path based DIDs.
-///
-/// # Errors
-///
-/// Returns a [DID resolution](https://www.w3.org/TR/did-core/#did-resolution-metadata)
-/// error as specified.
-pub async fn resolve(
-    did: &str, opts: Option<Options>, resolver: &impl DidResolver,
-) -> did::Result<Resolved> {
-    // use DID-specific resolver
-    let method = did.split(':').nth(1).unwrap_or_default();
-
-    let result = match method {
-        "key" => key::DidKey::resolve(did, opts, resolver),
-        "web" => web::DidWeb::resolve(did, opts, resolver).await,
-        _ => Err(Error::MethodNotSupported(format!("{method} is not supported"))),
-    };
-
-    if let Err(e) = result {
-        return Ok(Resolved {
-            metadata: Metadata {
-                error: Some(e.to_string()),
-                error_message: Some(e.message()),
-                content_type: ContentType::DidLdJson,
-                ..Metadata::default()
-            },
-            ..Resolved::default()
-        });
-    };
-
-    result
-}
-
 /// Dereference a DID URL into a resource.
 ///
 /// # Errors
 pub async fn dereference(
     did_url: &str, opts: Option<Options>, resolver: &impl DidResolver,
 ) -> did::Result<Dereferenced> {
-    let method = did_url.split(':').nth(1).unwrap_or_default();
+    // extract DID from DID URL
+    let url = url::Url::parse(did_url)
+        .map_err(|e| Error::InvalidDidUrl(format!("issue parsing URL: {e}")))?;
+    let did = format!("did:{}", url.path());
 
-    match method {
-        "key" => key::DidKey::dereference(did_url, opts, resolver),
-        "web" => web::DidWeb::dereference(did_url, opts, resolver).await,
-        _ => Err(Error::MethodNotSupported(format!("{method} is not supported"))),
-    }
+    // resolve DID document
+    let method = did_url.split(':').nth(1).unwrap_or_default();
+    let resolved = match method {
+        "key" => key::DidKey::resolve(&did, opts, resolver)?,
+        "web" => web::DidWeb::resolve(&did, opts, resolver).await?,
+        _ => return Err(Error::MethodNotSupported(format!("{method} is not supported"))),
+    };
+
+    let Some(document) = resolved.document else {
+        return Err(Error::InvalidDid("Unable to resolve DID document".into()));
+    };
+
+    // process document to dereference DID URL for requested resource
+    let Some(verifcation_methods) = document.verification_method else {
+        return Err(Error::NotFound("verification method missing".into()));
+    };
+
+    // for now we assume the DID URL is the ID of the verification method
+    // e.g. did:web:demo.credibil.io#key-0
+    let Some(vm) = verifcation_methods.iter().find(|vm| vm.id == did_url) else {
+        return Err(Error::NotFound("verification method not found".into()));
+    };
+
+    Ok(Dereferenced {
+        metadata: Metadata {
+            content_type: ContentType::DidLdJson,
+            ..Metadata::default()
+        },
+        content_stream: Some(Resource::VerificationMethod(vm.clone())),
+        content_metadata: Some(ContentMetadata {
+            document_metadata: resolved.document_metadata,
+        }),
+    })
 }
 
 /// Used to pass addtional values to a `resolve` and `dereference` methods. Any
@@ -108,45 +95,45 @@ pub struct Options {
     pub additional: Option<HashMap<String, Metadata>>,
 }
 
-/// The DID URL syntax supports parameters in the URL query component. Adding a DID
-/// parameter to a DID URL means the parameter becomes part of the identifier for a
-/// resource.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Parameters {
-    /// Identifies a service from the DID document by service's ID.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub service: Option<String>,
+// /// The DID URL syntax supports parameters in the URL query component. Adding a DID
+// /// parameter to a DID URL means the parameter becomes part of the identifier for a
+// /// resource.
+// #[derive(Clone, Debug, Default, Deserialize, Serialize)]
+// #[serde(rename_all = "camelCase")]
+// pub struct Parameters {
+//     /// Identifies a service from the DID document by service's ID.
+//     #[serde(skip_serializing_if = "Option::is_none")]
+//     pub service: Option<String>,
 
-    /// A relative URI reference that identifies a resource at a service endpoint,
-    /// which is selected from a DID document by using the service parameter.
-    /// MUST use URL encoding if set.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(alias = "relative-ref")]
-    pub relative_ref: Option<String>,
+//     /// A relative URI reference that identifies a resource at a service endpoint,
+//     /// which is selected from a DID document by using the service parameter.
+//     /// MUST use URL encoding if set.
+//     #[serde(skip_serializing_if = "Option::is_none")]
+//     #[serde(alias = "relative-ref")]
+//     pub relative_ref: Option<String>,
 
-    /// Identifies a specific version of a DID document to be resolved (the version ID
-    /// could be sequential, or a UUID, or method-specific).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub version_id: Option<String>,
+//     /// Identifies a specific version of a DID document to be resolved (the version ID
+//     /// could be sequential, or a UUID, or method-specific).
+//     #[serde(skip_serializing_if = "Option::is_none")]
+//     pub version_id: Option<String>,
 
-    /// Identifies a version timestamp of a DID document to be resolved. That is, the
-    /// DID document that was valid for a DID at a certain time.
-    /// An XML datetime value [XMLSCHEMA11-2] normalized to UTC 00:00:00 without
-    /// sub-second decimal precision. For example: 2020-12-20T19:17:47Z.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub version_time: Option<String>,
+//     /// Identifies a version timestamp of a DID document to be resolved. That is, the
+//     /// DID document that was valid for a DID at a certain time.
+//     /// An XML datetime value [XMLSCHEMA11-2] normalized to UTC 00:00:00 without
+//     /// sub-second decimal precision. For example: 2020-12-20T19:17:47Z.
+//     #[serde(skip_serializing_if = "Option::is_none")]
+//     pub version_time: Option<String>,
 
-    /// A resource hash of the DID document to add integrity protection, as specified
-    /// in [HASHLINK]. This parameter is non-normative.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(rename = "hl")]
-    pub hashlink: Option<String>,
+//     /// A resource hash of the DID document to add integrity protection, as specified
+//     /// in [HASHLINK]. This parameter is non-normative.
+//     #[serde(skip_serializing_if = "Option::is_none")]
+//     #[serde(rename = "hl")]
+//     pub hashlink: Option<String>,
 
-    /// Additional parameters.
-    #[serde(flatten)]
-    pub additional: Option<HashMap<String, Value>>,
-}
+//     /// Additional parameters.
+//     #[serde(flatten)]
+//     pub additional: Option<HashMap<String, Value>>,
+// }
 
 /// Returned by `resolve` DID methods.
 #[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
@@ -322,7 +309,19 @@ impl Error {
 
 #[cfg(test)]
 mod test {
+    use anyhow::anyhow;
+    use insta::assert_json_snapshot as assert_snapshot;
+
     use super::*;
+    use crate::Binding;
+
+    struct MockResolver;
+    impl DidResolver for MockResolver {
+        async fn resolve(&self, _binding: Binding) -> anyhow::Result<Document> {
+            serde_json::from_slice(include_bytes!("did/did-web-ecdsa.json"))
+                .map_err(|e| anyhow!("issue deserializing document: {e}"))
+        }
+    }
 
     #[test]
     fn error_code() {
@@ -330,5 +329,23 @@ mod test {
 
         println!("err: {:?}", err.to_string());
         assert_eq!(err.message(), "Method not supported");
+    }
+
+    #[tokio::test]
+    async fn deref_web() {
+        const DID_URL: &str = "did:web:demo.credibil.io#key-0";
+
+        let dereferenced =
+            dereference(DID_URL, None, &MockResolver).await.expect("should dereference");
+        assert_snapshot!("deref_web", dereferenced);
+    }
+
+    #[tokio::test]
+    async fn deref_key() {
+        const DID_URL: &str = "did:key:z6MkhaXgBZDvotDkL5257faiztiGiC2QtKLGpbnnEGta2doK#z6MkhaXgBZDvotDkL5257faiztiGiC2QtKLGpbnnEGta2doK";
+
+        let dereferenced =
+            dereference(DID_URL, None, &MockResolver).await.expect("should dereference");
+        assert_snapshot!("deref_key", dereferenced);
     }
 }
